feat: editSelectionContent에 enter, delete, write 통합 및 selectionChange를 changeSelectionStyle로 이름변경
